+++ @xen-orchestra/fs/src/xoSmb.js
@@ -1,43 +1,44 @@
 import execa from 'execa'
 import fs from 'fs-extra'
+import { join } from 'path'
 import { tmpdir } from 'os'
 
-import LocalHandler from './local'
+import MountHandler from './mountHandler'
 
-export default class SmbHandler extends LocalHandler {
-  constructor(remote, opts) {
+export default class SmbHandler extends MountHandler {
+  constructor(
+    remote,
+    { mountsDir = join(tmpdir(), 'xo-smb-mounts'), ...opts } = {}
+  ) {
     super(remote, opts)
 
     if (this._remote.path === '') {
-      this._remote.path = tmpdir() + '/smb/'
+      this._realPath = join(
+        mountsDir,
+        remote.id ||
+          Math.random()
+            .toString(36)
+            .slice(2)
+      )
+    } else {
+      this._realPath = this._remote.path
     }
-
-    const prefix = this._remote.path
-    this._prefix = prefix !== '' ? prefix + '\\' : prefix
   }
 
   get type() {
     return 'smb'
   }
 
-  async _forget() {
-    try {
-      await this._umount(this._remote)
-    } catch (error) {
-      throw error
-    }
-  }
-
   async _mount() {
-    const { host, path, username, password, domain } = this._remote
-    await fs.ensureDir(path)
+    await fs.ensureDir(this._getRealPath())
+    const { host, username, password, domain } = this._remote
     return execa(
       'mount',
       [
         '-t',
         'cifs',
         `\\\\${host}`,
-        path,
+        this._getRealPath(),
         '-o',
         `user=${username},password=${password},domain=${domain}`,
       ],
@@ -51,28 +52,6 @@ export default class SmbHandler extends LocalHandler {
         error == null ||
         typeof error.stderr !== 'string' ||
         !error.stderr.includes('already mounted')
-      ) {
-        throw error
-      }
-    })
-  }
-
-  async _sync() {
-    await this._mount()
-
-    return this._remote
-  }
-
-  async _umount() {
-    await execa('umount', ['--force', this._remote.path], {
-      env: {
-        LANG: 'C',
-      },
-    }).catch(error => {
-      if (
-        error == null ||
-        typeof error.stderr !== 'string' ||
-        !error.stderr.includes('not mounted')
       ) {
         throw error
       }
